Separate aws mapper completely from core
