+++ crates/extensions/tedge_mapper_core/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod component;
-pub mod converter;
 pub mod error;
 pub mod size_threshold;
